# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bridge-Watch API",
    "description": "Real-time Stellar bridge monitoring API. Provides bridge statuses, asset health, liquidity, price feeds, alert management, analytics, and administrative controls.\n\n## Authentication\nProtected endpoints require an API key in the `x-api-key` header.\n\n## Rate Limiting\nAll endpoints are rate-limited per IP using a sliding-window algorithm backed by Redis. Exceeding the limit returns **429 Too Many Requests** with a `Retry-After` header.\n\n## Versioning\nAll REST endpoints are prefixed with `/api/v1/`. Breaking changes increment the version segment.\n\n## Error Format\n```json\n{ \"error\": \"Short label\", \"message\": \"Human-readable description\" }\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Bridge-Watch Team",
      "url": "https://github.com/StellaBridge/Bridge-Watch"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" },
    { "url": "https://api.bridge-watch.io", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for protected endpoints"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/analytics/protocol": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get protocol-wide statistics",
        "parameters": [{ "name": "forceRefresh", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } }],
        "responses": { "200": { "description": "Protocol stats" } }
      }
    },
...
    "/api/v1/analytics/volume": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get volume aggregations",
        "parameters": [
          { "name": "period", "in": "query", "schema": { "type": "string", "enum": ["hourly", "daily", "weekly", "monthly"], "default": "daily" } },
          { "name": "symbol", "in": "query", "schema": { "type": "string" } },
          { "name": "bridgeName", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Volume aggregations" }, "400": { "description": "Invalid period" } }
      }
    },
    "/api/v1/analytics/trends/{metric}": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Calculate trend for a metric",
        "parameters": [{ "name": "metric", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Trend data" } }
      }
    },
    "/api/v1/analytics/top-performers": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get top performing assets or bridges",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["assets", "bridges"], "default": "assets" } },
          { "name": "metric", "in": "query", "schema": { "type": "string", "enum": ["volume", "tvl", "health"], "default": "health" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": { "200": { "description": "Top performers" } }
      }
    },
    "/api/v1/analytics/historical/{metric}": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get historical comparison data",
        "parameters": [
          { "name": "metric", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "symbol", "in": "query", "schema": { "type": "string" } },
          { "name": "days", "in": "query", "schema": { "type": "integer", "default": 30 } }
        ],
        "responses": { "200": { "description": "Historical data" } }
      }
    },
...
    "/api/v1/analytics/cache/invalidate": {
      "post": {
        "tags": ["Analytics"],
        "summary": "Invalidate analytics cache",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "pattern": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Cache invalidated" } }
      }
    },
...
    "/api/v1/cache/invalidate": {
      "post": {
        "tags": ["Cache"],
        "summary": "Invalidate cache entries",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "tag": { "type": "string" }, "key": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Invalidated" }, "400": { "description": "Missing tag or key" } }
      }
    },
...
  }
}